EV-266: Renamed Services to Service
