+++ src/Command/FeedImportCommand.php
@@ -4,8 +4,8 @@ namespace App\Command;
 
 use App\Message\FeedItemDataMessage;
 use App\Repository\FeedRepository;
-use App\Services\Feeds\Mapper\FeedConfigurationMapper;
-use App\Services\Feeds\Parser\FeedParserInterface;
+use App\Service\Feeds\Mapper\FeedConfigurationMapper;
+use App\Service\Feeds\Parser\FeedParserInterface;
 use CuyZ\Valinor\Mapper\MappingError;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
