+++ src/Command/FeedListCommand.php
@@ -3,7 +3,7 @@
 namespace App\Command;
 
 use App\Repository\FeedRepository;
-use App\Services\Feeds\Mapper\FeedConfigurationMapper;
+use App\Service\Feeds\Mapper\FeedConfigurationMapper;
 use CuyZ\Valinor\Mapper\MappingError;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
