+++ src/Service/ContentNormalizerInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Service;
 
 interface ContentNormalizerInterface
 {
