+++ src/Service/TagsNormalizer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Service;
 
 use App\Entity\Tag;
 use App\Repository\TagRepository;
